dogs_quiz/requirements.txt
Flask
gunicorn
pandas
numpy
Tensorflow-cpu
pillow
